home-manager: fix missing string context
`home-manager` package was being passed a `path` string pointing a
store path while the string didn't have context. As a result the
`home-manager` derivation was created without this path as an input
source allowing it to be garbage collected even though it it being used
by the `home-manager` script. This in turn caused certain warnings and
errors to occur.